Use inner state for displayed date in calendar
